edit balance with unlocking mechanism for it

+++ accounts/models.py
@@ -10,7 +10,7 @@ from users.models import User
 
 class Account(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='accounts')
-    
+
     ACCOUNT_TYPE_CHOICES = [
         ('Checking', 'Checking'),
         ('Savings', 'Savings'),
@@ -21,7 +21,7 @@ class Account(models.Model):
     slug = models.SlugField()
     name = models.CharField('Account Name', max_length=50)
     number = models.IntegerField('Account Number')
-    balance = models.FloatField('Balance', default=0)
+    balance = models.DecimalField('Balance', max_digits=1000, decimal_places=2, default=0)
     account_type = models.CharField('Type', max_length=10, choices=ACCOUNT_TYPE_CHOICES)
     bank = models.CharField('Bank', max_length=20, null=True)
     notes = models.TextField(null=True)
